[improve_computation_time] Fixed implicit casts
Problems detected by clang (-Wconversion works better)

+++ code/hydro_models/src/WaveModelInterface.cpp
@@ -40,9 +40,9 @@ double WaveModelInterface::get_relative_wave_height(const Point& P, const TR1(sh
 std::vector<double> WaveModelInterface::get_relative_wave_height(const PointMatrix& P, const TR1(shared_ptr)<Kinematics>& k) const
 {
     const PointMatrix OP = compute_relative_position(P, k);
-    const size_t n = P.m.cols();
+    const int n = P.m.cols();
     std::vector<double> ret;
-    for (size_t i = 0 ; i < n ; ++i)
+    for (int i = 0 ; i < n ; ++i)
     {
         ret.push_back(wave_height(OP.m(0,i),OP.m(1,i),OP.m(2,i)));
     }
